ml-training/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
paddlepaddle==2.5.2
paddleocr==2.7.3
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
tqdm==4.66.1
boto3==1.29.7
pytest==7.4.3
pytest-mock==3.12.0